Convert `auth/confirmations` spec controller->request (#37735)
